Refactor wasm bindings into SRP modules
