+++ crates/tokmd-wasm/src/error.rs
@@ -1,0 +1,6 @@
+use js_sys::Error as JsError;
+use wasm_bindgen::prelude::*;
+
+pub(crate) fn to_js_error(message: impl Into<String>) -> JsValue {
+    JsError::new(&message.into()).into()
+}
